requirements.txt
django>=5.2
djangorestframework>=3.16
psycopg2-binary>=2.9
